requirements-drac.txt
accelerate
hydra-core >=1.3.0
hydra-submitit-launcher
matplotlib
omegaconf >=2.1.0
peft
scikit-learn
seaborn
submitit
torch >=2.3.0
tqdm >=4.66.5
transformers >=4.40.0,<4.48